Issues Data Table: Move TABLE_COLUMNS enum to another module
There is circular dependency warning being raised.

Let's move TABLE_COLUMNS enum into another module to fix this.
