+++ src/app/phase-team-response/issues-faulty/issues-faulty.component.ts
@@ -5,7 +5,8 @@ import { Issue } from '../../core/models/issue.model';
 import { UserService } from '../../core/services/user.service';
 import { UserRole } from '../../core/models/user.model';
 import { PermissionService } from '../../core/services/permission.service';
-import { ACTION_BUTTONS, IssueTablesComponent, TABLE_COLUMNS } from '../../shared/issue-tables/issue-tables.component';
+import { ACTION_BUTTONS, IssueTablesComponent } from '../../shared/issue-tables/issue-tables.component';
+import { TABLE_COLUMNS } from '../../shared/issue-tables/issue-tables-columns';
 
 @Component({
   selector: 'app-issues-faulty',
