+++ src/app/phase-tester-response/issue-responded/issue-responded.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { ACTION_BUTTONS, IssueTablesComponent, TABLE_COLUMNS } from '../../shared/issue-tables/issue-tables.component';
+import { ACTION_BUTTONS, IssueTablesComponent } from '../../shared/issue-tables/issue-tables.component';
+import { TABLE_COLUMNS } from '../../shared/issue-tables/issue-tables-columns';
 import { Issue, STATUS } from '../../core/models/issue.model';
 
 @Component({
