+++ src/app/shared/issue-tables/issue-tables-columns.ts
@@ -1,0 +1,12 @@
+export enum TABLE_COLUMNS {
+    ID = 'id',
+    TITLE = 'title',
+    TEAM_ASSIGNED = 'teamAssigned',
+    TYPE = 'type',
+    SEVERITY = 'severity',
+    RESPONSE = 'responseTag',
+    ASSIGNEE = 'assignees',
+    DUPLICATED_ISSUES = 'duplicatedIssues',
+    TODO = 'Todo Remaining',
+    ACTIONS = 'actions'
+  }
