+++ src/app/shared/issue-tables/issue-tables.component.ts
@@ -20,19 +20,6 @@ export enum ACTION_BUTTONS {
   DELETE_ISSUE
 }
 
-export enum TABLE_COLUMNS {
-  ID = 'id',
-  TITLE = 'title',
-  TEAM_ASSIGNED = 'teamAssigned',
-  TYPE = 'type',
-  SEVERITY = 'severity',
-  RESPONSE = 'responseTag',
-  ASSIGNEE = 'assignees',
-  DUPLICATED_ISSUES = 'duplicatedIssues',
-  TODO = 'Todo Remaining',
-  ACTIONS = 'actions'
-}
-
 @Component({
   selector: 'app-issue-tables',
   templateUrl: './issue-tables.component.html',
